+++ test/testIsPartOf.ts
@@ -3,7 +3,7 @@ import * as assert from 'assert';
 import * as sriClientFactory from '@kathondvla/sri-client/node-sri-client';
 import utilsFactory from './utils';
 
-export = module.exports = function (base) {
+module.exports = function (base) {
   const sriClientConfig = {
     baseUrl: base,
   };