+++ test/testOrderBy.ts
@@ -2,7 +2,7 @@
 var assert = require('assert');
 import * as sriclientFactory from '@kathondvla/sri-client/node-sri-client';
 
-export = module.exports = function (base) {
+module.exports = function (base) {
 
 
   const sriClientConfig = {